% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDT.R
\name{PDT}
\alias{PDT}
\title{Launch the Pitch Discrimination Task (PDT)}
\usage{
PDT(
  num_items = 10L,
  with_final_page = TRUE,
  headphones_page = TRUE,
  import_musicassessr_js_scripts = TRUE
)
}
\arguments{
\item{import_musicassessr_js_scripts}{}
}
\value{

}
\description{
Launch the Pitch Discrimination Task (PDT)
}
