% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDT.R
\name{PDT_standalone}
\alias{PDT_standalone}
\title{Launch PDT as a standalone test}
\usage{
PDT_standalone(
  title = "Pitch Discrimination Task",
  admin_password = "demo",
  researcher_email = "sebsilas@gmail.com",
  num_items = 10L
)
}
\arguments{
\item{num_items}{}
}
\value{

}
\description{
Launch PDT as a standalone test
}
