% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDT.R
\name{PDT}
\alias{PDT}
\title{Launch the Pitch Discrimination Task (PDT)}
\usage{
PDT(num_items = 10L, with_final_page = TRUE)
}
\arguments{
\item{with_final_page}{}
}
\value{

}
\description{
Launch the Pitch Discrimination Task (PDT)
}
